+++ Vertex.java
@@ -2,11 +2,11 @@ import java.util.*;
 
 public class Vertex {
     public Point3d p;
-    public double value;
+    public float value;
     public int index;
     public Set<Vertex> neighbours = new HashSet<>();
 
-    public Vertex(Point3d p, double value, int index) {
+    public Vertex(Point3d p, float value, int index) {
         this.p = p;
         this.value = value;
         this.index = index;
